refacto: dry things up in the default grade tallies
